Move hash to utils, add comments
